refactor(sonic_similarity): extract pure helpers and dataclasses
Second stage of the split: moves _parse_clap_embedding, _parse_weights,
_parse_similar_params, and apply_filters to helpers.py, and the
SimilarParams + _SearchContext dataclasses to models.py. No behavior
change. test_plugin_api imports updated to source the helpers from
.helpers directly. __init__.py shrinks by another 150 lines (1692 → 1542).
